backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-jose
passlib[bcrypt]
python-multipart
PyYAML
